backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
anthropic==0.28.0
replicate==0.30.1
supabase==2.5.3
numpy==1.26.4
scikit-image==0.23.2
colormath==3.0.0
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
